extract into method
