+++ backend/src/nodes/nodes/image/paired_image_file_iterator.py
@@ -19,7 +19,7 @@ from ...properties.outputs import (
     NumberOutput,
 )
 from ...utils.image_utils import get_available_image_formats
-from ...utils.utils import alphanumeric_sort
+from ...utils.utils import alphanumeric_sort, walk_sorted
 
 PAIRED_IMAGE_ITERATOR_NODE_ID = "chainner:image:paired_file_iterator_load"
 
@@ -105,36 +105,8 @@ class PairedImageFileIteratorNode(IteratorNodeBase):
 
         supported_filetypes = get_available_image_formats()
 
-        def walk_error_handler(exception_instance):
-            logger.warning(
-                f"Exception occurred during walk: {exception_instance} Continuing..."
-            )
-
-        just_image_files_a: List[str] = []
-        for root, dirs, files in os.walk(
-            directory_a, topdown=True, onerror=walk_error_handler
-        ):
-            await context.progress.suspend()
-
-            dirs.sort(key=alphanumeric_sort)
-            for name in sorted(files, key=alphanumeric_sort):
-                filepath = os.path.join(root, name)
-                _base, ext = os.path.splitext(filepath)
-                if ext.lower() in supported_filetypes:
-                    just_image_files_a.append(filepath)
-
-        just_image_files_b: List[str] = []
-        for root, dirs, files in os.walk(
-            directory_b, topdown=True, onerror=walk_error_handler
-        ):
-            await context.progress.suspend()
-
-            dirs.sort(key=alphanumeric_sort)
-            for name in sorted(files, key=alphanumeric_sort):
-                filepath = os.path.join(root, name)
-                _base, ext = os.path.splitext(filepath)
-                if ext.lower() in supported_filetypes:
-                    just_image_files_b.append(filepath)
+        just_image_files_a: List[str] = walk_sorted(directory_a, supported_filetypes)
+        just_image_files_b: List[str] = walk_sorted(directory_b, supported_filetypes)
 
         assert len(just_image_files_a) == len(just_image_files_b), (
             "Number of images in directories A and B must be equal. "
